Setting up a utility file for testing and otherwise implementing more tests based on the new structure
